ISSUE-490 refactored, fixed build errors
